refactor(domain): PrizeReception보다 더 명확한 이름인 ProfitCalculator로 변경하고 사용하지 않는 메서드 제거
